Split drawing.rs into multiple files. Doesn't change public API
